Add files via upload
ASSIGNMENT OVERVIEW
This assignment will allow you to experience what it is like to develop a complete website from scratch.  To do this you will create a site of your choice.  When you are done there will be five (5) pages – a Template, Home, Image, Media and Feedback pages. (Template will not be linked to the other pages) Instructions on specific requirements are as follows;