Add files via upload
ASSIGNMENT OVERVIEW
Create a C# console program that creates and records information for a T-Shirt store. To manage the size of inventory, the owners only want to store data for a total of 12 T-Shirts. Assume the store can only hold three T-Shirt brands with one of four sizes for each brand which are Small (1), Medium (2), Large (3), XLarge (4).

When the program is first run get the name of each of the three brands and store in an array called “brand+<yourGroupInitials>” for example brandJSBAMW. Use a for loop to perform this operation.


the following information must be displayed in a menu:

Add new T-Shirt details -Brand Name & Size (for example Fendi-4)

Edit existing T-Shirt details (Brand Name & Size)

Display all T-Shirts in store (display the Brand Name & Size)

Delete T-Shirt Information

Exit the program

After a choice from the main menu is selected, the program must accurately perform the necessary operations and returns to the main menu unless option "E" is selected.