Add files via upload
ASSIGNMENT OVERVIEW
Create a web page that allows a user to take a quiz with 5 questions.

There should be at least one question with True/False, one question where the answer is a number and one question where the answer is a word/string. The other 2 questions can be any of these types as you choose.

User should be able to attempt a question by clicking a button for each question.

Once the user attempts a question, it should update the user whether the answer is correct/incorrect by displaying the word "Correct" or "Incorrect" in front of/near the button used to attempt that question. The colour of the word "Correct" should be different from the colour of the word "Incorrect". The colours are your choice.

After each question is attempted, the total score on the page should be updated.

For the question, where the answer is a numeric value, if the user inputs some value that is not a number, the program should notify the user and ask for a numeric value again as long as the user does not enter a numeric value.

I have attached a few screenshots that may help you understand the tasks better.